I believe i fixed the errors on my end

+++ PolygonChecker/main.c
@@ -17,9 +17,9 @@ int main() {
 		{
 		case 1:
 			printf_s("Triangle selected.\n");
-			int triangleSides[3] = { 0, 0, 0 };
-			int* triangleSidesPtr = getTriangleSides(triangleSides);
-			//printf_s("! %d\n", triangleSidesPtr[0]);
+			double triangleSides[3] = { 0, 0, 0 };
+			double* triangleSidesPtr = getTriangleSides(triangleSides);
+
 			char* result = analyzeTriangle(triangleSidesPtr[0], triangleSidesPtr[1], triangleSidesPtr[2]);
 			printf_s("%s\n", result);
 			break;
@@ -29,7 +29,7 @@ int main() {
 			double rectanglePoint[4][2];
 			getrectangleSides(rectanglePoint);
 			for (int i = 0; i < 4; i++) {
-				printf("Side %d is: [%d, %d] ", i + 1, rectanglePoint[i][0], rectanglePoint[i][1]);
+				printf("Side %d is: [%lf, %lf] ", i + 1, rectanglePoint[i][0], rectanglePoint[i][1]);
 			}
 			break;
 
@@ -67,11 +67,11 @@ int printShapeMenu() {
 	return shapeChoice;
 }
 
-int* getTriangleSides(int* triangleSides) {
+double* getTriangleSides(double* triangleSides) {
 	printf_s("Enter the three sides of the triangle: ");
 	for (int i = 0; i < 3; i++)
 	{
-		scanf_s("%d", &triangleSides[i]);
+		scanf_s("%lf", &triangleSides[i]);
 	}
 	return triangleSides;
 }
@@ -83,7 +83,7 @@ void getrectangleSides(double points[4][2]) {
 		while(recSidesValid == true)
 		{
 			printf_s("Please enter the points you want 1 at a time, in this format(x y):\n");
-			if (scanf_s("%d %d", &points[i][0], &points[i][1]) == 2) {
+			if (scanf_s("%lf %lf", &points[i][0], &points[i][1]) == 2) {
 				break;
 			}
